simplify isValidSumFormula to not use getColAndRow, move currently unused code to old-code
