+++ old-code/is-sum-function-test.js
